Added node weight support
Note that weight can be specified using the node-modify only but will
be displayed in load balancer status, node-list, node-status, node-add,
and node-modify. To support the setting of weight with load balancer
create or node-add, it would require a bigger change because with those,
multiple nodes can be specified with --node.  Perhaps a syntax of
ip:port:weight would make sense if that was supported.

Also note that the API interface only returns weight when the value is
not the default (1).  Therefore, the weight column also only displays
a weight value when it is not 1.

Change-Id: I34ea3e4a144aeb47af661c161f5f3ee4946cbe7e